pyproject.toml
[project]
name = "multiagents-travelplanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "chromadb>=0.4.24",
    "docx>=0.2.4",
    "fpdf==1.7.2",
    "ipython>=9.5.0",
    "langchain-chroma>=0.1.4",
    "langchain-community>=0.2.19",
    "langchain-openai>=0.1.25",
    "langchain-text-splitters>=0.2.4",
    "langgraph==0.2.14",
    "matplotlib>=3.10.6",
    "pillow>=10.4.0",
    "pyprojroot>=0.3.0",
    "python-docx>=1.2.0",
    "python-dotenv==1.0.1",
    "requests>=2.31.0",
    "sentence-transformers>=2.5.0",
    "streamlit==1.38.0",
    "unstructured[docx]>=0.18.14",
]
